% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_cone_radius.R
\name{benchmark_cone_radius}
\alias{benchmark_cone_radius}
\title{Benchmark Different Implementations of cone_radius}
\arguments{
\item{implementations}{A named list of functions to benchmark. Defaults to:
\itemize{
\item \code{For_Loop}: \code{cone_radius_for}
\item \code{Purrr_Map}: \code{cone_radius_map}
\item \code{Apply}: \code{cone_radius_apply}
\item \code{Direct_Vectorized}: \code{cone_radius_vectorized_direct}
\item \code{Rcpp}: \code{cone_radius_cpp}
}}

\item{x_values}{A numeric vector representing the heights to test
(default: seq(0, 10, length.out = 1000)).}

\item{times}{An integer specifying the number of iterations for the benchmark
(default: 100).}

\item{plot}{A logical value indicating whether to generate a plot of the results
(default: FALSE).}
}
\value{
A microbenchmark object containing the benchmark results.
}
\description{
Benchmarks various implementations of the \code{cone_radius} function
for a predefined sequence of values and number of iterations. By default,
benchmarks all implementations on a sequence of 1000 values over 100 iterations.
}
\examples{
# Default benchmark
benchmark_cone_radius()

# Custom benchmark
benchmark_cone_radius(
  implementations = list(
    For_Loop = cone_radius_for,
    Rcpp = cone_radius_cpp
  ),
  x_values = seq(0, 10, length.out = 500),
  times = 50,
  plot = TRUE
)
}
